% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_point_pie.R
\docType{data}
\name{StatPointPie}
\alias{StatPointPie}
\title{StatPointPie}
\format{
An object of class \code{StatPointPie} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatPointPie
}
\description{
StatPointPie
}
\details{
Rethink about the design. Use this stat
to transform the "long" format data (one observation per row)
into the proper plot format, i.e, a piece of the pie per row.

This probably allow user to supply different data input, and they can
just choose different stat when call geom...

Use StatPointPie to turn long format to short format:
group by \code{fill} + either \code{group} or \verb{x,y},
count the number of observations.

if \code{group} is supplied, Stat will recieve each group
}
\keyword{datasets}
