% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_point_pie.R, R/stat_point_pie.R
\docType{data}
\name{ggproto-subclass}
\alias{ggproto-subclass}
\alias{GeomPointPie}
\alias{StatPointPie}
\title{PointPie ggproto classes}
\format{
An object of class \code{GeomPointPie} (inherits from \code{GeomPolygon}, \code{Geom}, \code{ggproto}, \code{gg}) of length 5.

An object of class \code{StatPointPie} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
GeomPointPie

StatPointPie
}
\description{
\code{\link[=ggproto]{ggproto()}} class that describes combination of pie chart and point.
}
\details{
This is the subclass inherited from \code{\link[=ggproto]{ggproto()}}. Use the main layer functions
\code{\link[=geom_point_pie]{geom_point_pie()}} instead of this object.

The stat to calculate the count of each subgroup to handle the long data
format
}
\note{
amount Useful when the user calculate the amount for each subgroup.

subgroup Each complete pie chart
}
\keyword{datasets}
