% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_pie.R
\name{geom_point_pie}
\alias{geom_point_pie}
\title{Draw pie chart at any x,y coordinates.}
\usage{
geom_point_pie(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  r0 = 0L,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{r0}{The radius (0-1) of the inner circle, in case a dounut plot is needed.}
}
\description{
This geom is based on \code{ggplot2::GeomPolygon}. We used polygon geom to create pie
chart than can be mapped to arbitrary x and y coordinates
}
\details{
Aesthetics:
geom_point_pie understand the following aesthetics (required aesthetics are in
bold):
\itemize{
\item \strong{x}
\item \strong{y}
\item \strong{group}
\item color
\item fill
\item size
\item linetype (will change everything rn)
\item alpha
}
}
