% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_pie.R
\docType{data}
\name{GeomPointPie}
\alias{GeomPointPie}
\title{PointPieGeom Proto}
\format{
An object of class \code{GeomPointPie} (inherits from \code{GeomPolygon}, \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomPointPie
}
\arguments{
\item{...}{skip for now}
}
\description{
ggproto class that descripe combination of pie chart and point.

ggplot always map size to 1-6. Use scale_size_contineuous to deal with it.
In the future, maybe get our own scale...
}
\keyword{datasets}
