% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_pie.R
\docType{data}
\name{GeomPointPie}
\alias{GeomPointPie}
\title{I think the r1  and r0 are in their own system?}
\format{
An object of class \code{GeomPointPie} (inherits from \code{GeomPolygon}, \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomPointPie
}
\description{
I think the r1  and r0 are in their own system?
}
\keyword{datasets}
