% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_pie.R
\name{pieGrob}
\alias{pieGrob}
\title{Grob to handle a single pie chart.}
\usage{
pieGrob(
  x = 0.5,
  y = 0.5,
  r0 = 0.1,
  r1 = 0.3,
  theta0 = 0,
  theta1 = pi/2,
  n = 360,
  name = NULL,
  default.units = "npc",
  gp = gpar(),
  vp = NULL
)
}
\description{
Based on grid::polygonGrob.
Sample points along the arc to create a pseudo-circle.
thetas are assumed to be in radius, not degree/
}
\details{
Assume x/y coordinate and radius are in the same coordinate system.
In other words, I simplyly added those two values.
When building geom, maybe need to looking into how to deal with the size
and x/y coords.
}
