% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-r1.R
\name{scale_r1}
\alias{scale_r1}
\title{The scales to map the point pie radius.}
\usage{
scale_r1(..., range = c(0.1, 0.4))
}
\description{
These scales are used for the \code{r1} aes in \link{geom_point_pie} layer as default.
We currently do not support any discrete scale.
}
